+++ src/gameEngine/baccaratGameEngine.js
@@ -26,6 +26,7 @@ class BaccaratGameEngine {
      */
     burnCards() {
         var burnCard = this.shoe.draw();
+        var burnCards = [];
 
         var burnCardValue = BaccaratResultsEngine.valueForCard(burnCard);
 
@@ -34,10 +35,10 @@ class BaccaratGameEngine {
             burnCardValue = 10;
 
         for (var i = 0; i < burnCardValue; i++) {
-            this.shoe.draw();
+            burnCards.push(this.shoe.draw());
         }
 
-        return burnCard;
+        return {burnCard, burnCards};
     }
     
     dealGame() {
@@ -72,23 +73,50 @@ class BaccaratGameEngine {
             var player3rdCard = this.shoe.draw();
             hand.playerCards.push(player3rdCard);
             var player3rdCardValue =  BaccaratResultsEngine.valueForCard(player3rdCard);
-            playerCardsValue = this.resultsEngine.calculateHandValue(hand.playerCards);
 
-            if (bankerCardsValue <= 2) {
-                bankerDraw = true;
+            switch (player3rdCardValue) {
+                case 2:
+                case 3:
+                // Player has 2, 3 - banker draws 0-4, stands 5-7
+                if (bankerCardsValue < 5) bankerDraw = true;
+                break;
+                case 4:
+                case 5:
+                // Player has 4, 5 - banker draws 0-5, stands 6-7
+                if (bankerCardsValue < 6) bankerDraw = true;
+                break;
+                case 6:
+                case 7:
+                // Player has 6, 7 - banker draws 0-6, stands 7
+                if (bankerCardsValue < 7) bankerDraw = true;
+                break;
+                case 8:
+                // Player has 8 - banker draws 0-2, stands 3-7
+                if (bankerCardsValue < 3) bankerDraw = true;
+                break;
+                case 9:
+                case 0:
+                case 1:
+                // Player has 9, T/K/Q/J, A - banker draws 0-3, stands 4-7
+                if (bankerCardsValue < 4) bankerDraw = true;
+                break;
             }
-            else if (bankerCardsValue == 3 && player3rdCardValue != 8)  {
-                bankerDraw = true;
-            }
-            else if (bankerCardsValue == 4 && player3rdCardValue >= 2 && player3rdCardValue <= 7) {
-                bankerDraw = true;
-            }
-            else if (bankerCardsValue == 5 && player3rdCardValue >= 4 && player3rdCardValue <= 7) {
-                bankerDraw = true;
-            }
-            else if (bankerCardsValue == 6 && playerCardsValue >= 6 && playerCardsValue <= 7) {
-                bankerDraw = true;
-            }
+
+            // if (bankerCardsValue <= 2) {
+            //     bankerDraw = true;
+            // }
+            // else if (bankerCardsValue == 3 && player3rdCardValue != 8)  {
+            //     bankerDraw = true;
+            // }
+            // else if (bankerCardsValue == 4 && player3rdCardValue >= 2 && player3rdCardValue <= 7) {
+            //     bankerDraw = true;
+            // }
+            // else if (bankerCardsValue == 5 && player3rdCardValue >= 4 && player3rdCardValue <= 7) {
+            //     bankerDraw = true;
+            // }
+            // else if (bankerCardsValue == 6 && playerCardsValue >= 6 && playerCardsValue <= 7) {
+            //     bankerDraw = true;
+            // }
         }
 
         if (bankerDraw) {