+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.3 2020-07-04]
+### Added
+### Changed
+- Incorrect drawing rules in some cases for banker.
+- Return the individual burn cards in API
+
+
 ## [1.1.2 2020-04-15]
 ### Added
 - Cut cards to the shoe implementation